+++ tests/conftest.py
@@ -1,8 +1,16 @@
+from typing import List
+
 import pytest
 import requests
 
+from dicomtrolley.core import Series, Study
+from dicomtrolley.dicom_qr import DICOMQR
 from dicomtrolley.mint import Mint, MintStudy, parse_mint_studies_response
 from dicomtrolley.wado import Wado
+from tests.factories import (
+    create_c_find_image_response,
+    create_c_find_study_response,
+)
 from tests.mock_responses import (
     LOGIN_DENIED,
     LOGIN_SUCCESS,
@@ -53,25 +61,79 @@ def a_wado(a_session):
 
 
 @pytest.fixture
-def a_study_with_instances() -> MintStudy:
+def a_mint_study_with_instances() -> MintStudy:
     """An example MintStudy object"""
     studies = parse_mint_studies_response(MINT_SEARCH_INSTANCE_LEVEL.text)
     return studies[0]
 
 
 @pytest.fixture
-def a_study_without_instances() -> MintStudy:
+def a_mint_study_without_instances() -> MintStudy:
     """An example MintStudy object"""
     studies = parse_mint_studies_response(MINT_SEARCH_STUDY_LEVEL.text)
     return studies[0]
 
 
 @pytest.fixture
-def some_studies(a_study_with_instances, a_study_without_instances):
-    return [a_study_with_instances, a_study_with_instances]
+def some_mint_studies(
+    a_mint_study_with_instances, a_mint_study_without_instances
+):
+    return [a_mint_study_with_instances, a_mint_study_without_instances]
 
 
 def set_mock_response(requests_mock, response):
     """Register the given MockResponse with requests_mock"""
     requests_mock.register_uri(**response.as_dict())
     return response
+
+
+@pytest.fixture
+def an_image_level_study() -> List[Study]:
+    """A study with series and slice info"""
+    response = create_c_find_image_response(
+        study_instance_uid="Study1",
+        series_instance_uids=["Series1", "Series2"],
+        sop_class_uids=[f"Instance{i}" for i in range(1, 10)],
+    )
+    return DICOMQR.parse_c_find_response(response)
+
+
+@pytest.fixture
+def an_image_level_series() -> Series:
+    """A study with series and slice info"""
+    response = create_c_find_image_response(
+        study_instance_uid="Study1",
+        series_instance_uids=["Series1"],
+        sop_class_uids=[f"Instance{i}" for i in range(1, 10)],
+    )
+    study = DICOMQR.parse_c_find_response(response)[0]
+    return study.get("Series1")
+
+
+@pytest.fixture
+def another_image_level_series() -> Series:
+    """A study with series and slice info"""
+    response = create_c_find_image_response(
+        study_instance_uid="Study1",
+        series_instance_uids=["Series2"],
+        sop_class_uids=[f"Instance{i}" for i in range(1, 10)],
+    )
+    study = DICOMQR.parse_c_find_response(response)[0]
+    return study.get("Series2")
+
+
+@pytest.fixture
+def a_study_level_study():
+    """Study witnout slice info"""
+    return DICOMQR.parse_c_find_response(
+        create_c_find_study_response(study_instance_uids=["Study2"])
+    )
+
+
+@pytest.fixture
+def some_studies(an_image_level_study, a_study_level_study):
+    """Two studies. One at image level, with all slice info included. One at
+    study level, without slice info
+    """
+
+    return an_image_level_study + a_study_level_study